middleware/requirements.txt
Flask
torch
torchvision
nltk
google-generativeai
transformers
python-dotenv
flask-cors
PyMuPDF
sentencepiece
apscheduler